+++ packages/angular/src/client/vcd.api.client.ts
@@ -468,14 +468,14 @@ export class VcdApiClient {
         this._cloudApiSessionLinks.next(links || []);
     }
 
-    public get<T>(endpoint: string): Observable<T> {
+    public get<T>(endpoint: string,  options?: { headers?: HttpHeaders }): Observable<T> {
         return this.validateRequestContext().pipe(
-            concatMap(() => this.http.get<T>(`${this._baseUrl}/${endpoint}`))
+            concatMap(() => this.http.get<T>(this.buildEndpointUrl(endpoint), {...options}))
         );
     }
 
     public list<T>(endpoint: string, queryBuilder?: Query.Builder, multisite?: boolean | AuthorizedLocationType[]) {
-        let url = `${this._baseUrl}/${endpoint}`;
+        let url = this.buildEndpointUrl(endpoint);
 
         if (queryBuilder) {
             url = `${url}${queryBuilder.getCloudAPI()}`;
@@ -487,22 +487,30 @@ export class VcdApiClient {
         );
     }
 
-    public createSync<T>(endpoint: string, item: T): Observable<T> {
+    public createSync<T>(endpoint: string, item: T, options?: { headers?: HttpHeaders }): Observable<T> {
         return this.validateRequestContext().pipe(
-            concatMap(() => this.http.post<T>(`${this._baseUrl}/${endpoint}`, item))
+            concatMap(() => this.http.post<T>(
+                this.buildEndpointUrl(endpoint),
+                item,
+                {...options}
+            ))
         );
     }
 
-    public createAsync<T>(endpoint: string, item: T): Observable<TaskType> {
+    public createAsync<T>(endpoint: string, item: T, options?: { headers?: HttpHeaders }): Observable<TaskType> {
         return this.validateRequestContext().pipe(
-            concatMap(() => this.http.post(`${this._baseUrl}/${endpoint}`, item, { observe: 'response' })),
+            concatMap(() => this.http.post(
+                this.buildEndpointUrl(endpoint),
+                item,
+                { ...options, observe: 'response' }
+            )),
             concatMap(response => this.mapResponseToTask(response, 'POST'))
         );
     }
 
     public getTransferLink<T>(endpoint: string, item: T, transferRel: string = TRANSFER_LINK_REL): Observable<string> {
         return this.http
-            .post(`${this._baseUrl}/${endpoint}`, item, { observe: 'response' })
+            .post(this.buildEndpointUrl(endpoint), item, { observe: 'response' })
             .pipe(
                 map((res: HttpResponse<T & Navigable>) => {
                     const headerLinks: LinkType[] = res.headers.has(HATEOAS_HEADER)
@@ -526,42 +534,37 @@ export class VcdApiClient {
             );
     }
 
-    public updateSync<T>(endpoint: string, item: T, options?: {
-        headers?: HttpHeaders
-    }): Observable<T> {
+    public updateSync<T>(endpoint: string, item: T, options?: { headers?: HttpHeaders }): Observable<T> {
         return this.validateRequestContext().pipe(
             concatMap(() => this.http.put<T>(
-                `${this._baseUrl}/${endpoint}`,
+                this.buildEndpointUrl(endpoint),
                 item,
-                {
-                    ...options
-                }
+                { ...options }
             ))
         );
     }
 
-    public updateAsync<T>(endpoint: string, item: T, options?: {
-        headers?: HttpHeaders
-    }): Observable<TaskType> {
+    public updateAsync<T>(endpoint: string, item: T, options?: { headers?: HttpHeaders }): Observable<TaskType> {
 
         return this.validateRequestContext().pipe(
-            concatMap(() => this.http.put(`${this._baseUrl}/${endpoint}`, item, {
-                observe: 'response',
-                ...options
-            })),
+            concatMap(() => this.http.put(
+                this.buildEndpointUrl(endpoint),
+                item,
+                { ...options, observe: 'response' }
+            )),
             concatMap(response => this.mapResponseToTask(response, 'PUT'))
         );
     }
 
-    public deleteSync(endpoint: string): Observable<void> {
+    public deleteSync(endpoint: string, options?: { headers?: HttpHeaders }): Observable<void> {
         return this.validateRequestContext().pipe(
-            concatMap(() => this.http.delete<void>(`${this._baseUrl}/${endpoint}`))
+            concatMap(() => this.http.delete<void>(this.buildEndpointUrl(endpoint), { ...options }))
         );
     }
 
-    public deleteAsync(endpoint: string): Observable<TaskType> {
+    public deleteAsync(endpoint: string, options?: { headers?: HttpHeaders }): Observable<TaskType> {
         return this.validateRequestContext().pipe(
-            concatMap(() => this.http.delete(`${this._baseUrl}/${endpoint}`, { observe: 'response' })),
+            concatMap(() => this.http.delete(this.buildEndpointUrl(endpoint), { ...options, observe: 'response' })),
             concatMap(response => this.mapResponseToTask(response, 'DELETE'))
         );
     }
@@ -886,4 +889,13 @@ export class VcdApiClient {
     public getLocation(session: SessionType): AuthorizedLocationType {
         return session.authorizedLocations.location.find(location => location.locationId === session.locationId);
     }
+
+    /**
+     * Build the endpoint url. If the provided endpoint is already an absolute URL, then return it as it is without
+     * any modifications, otherwise consider it as a relative one and prepend the baseUrl as defined by the host application.
+     */
+    private buildEndpointUrl(endpoint: string): string {
+        return endpoint.indexOf('://') > -1 ? endpoint : `${this._baseUrl}/${endpoint}`;
+    }
+
 }